# docs/openapi.v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SHCW2026 OpenClaw API V2",
    "version": "2.2.0",
    "description": "Machine-readable OpenAPI spec for SHCW2026 /api/v2 endpoints. V2 preserves V1 behavior and adds response header X-API-Version plus standardized error code field, event registration audience endpoints, and expanded user management operations."
  },
  "servers": [
    {
      "url": "https://shcw2026.vercel.app",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "xApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v2/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List users",
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "in": "query", "name": "pageSize", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } },
          { "in": "query", "name": "search", "schema": { "type": "string" } },
          { "in": "query", "name": "role", "schema": { "type": "string" } },
          { "in": "query", "name": "status", "schema": { "type": "string" } },
          { "in": "query", "name": "sortBy", "schema": { "type": "string" } },
          { "in": "query", "name": "sortOrder", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" } }
        ],
        "responses": {
          "200": { "description": "Users list", "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaginatedSuccessResponse" } } } },
          "401": { "$ref": "#/components/responses/ErrorUnauthorized" },
          "403": { "$ref": "#/components/responses/ErrorForbidden" },
          "500": { "$ref": "#/components/responses/ErrorInternal" }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 },
                  "role": { "type": "string" },
                  "status": { "type": "string" },
                  "phone": { "type": "string" },
                  "title": { "type": "string" },
                  "bio": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created", "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "400": { "$ref": "#/components/responses/ErrorBadRequest" },
          "401": { "$ref": "#/components/responses/ErrorUnauthorized" },
          "403": { "$ref": "#/components/responses/ErrorForbidden" },
          "409": { "$ref": "#/components/responses/ErrorConflict" },
          "500": { "$ref": "#/components/responses/ErrorInternal" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Authorization: Bearer sk_oc_..."
      },
      "xApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "x-api-key: sk_oc_..."
      }
    },
    "headers": {
      "XApiVersion": {
        "description": "API version marker",
        "schema": {
          "type": "string",
          "enum": ["v2"]
        }
      }
    },
    "responses": {
      "ErrorBadRequest": {
        "description": "Bad Request",
        "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "success": false, "error": "Invalid request", "code": "BAD_REQUEST" } } }
      },
      "ErrorUnauthorized": {
        "description": "Unauthorized",
        "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "success": false, "error": "Unauthorized", "code": "UNAUTHORIZED" } } }
      },
      "ErrorForbidden": {
        "description": "Forbidden",
        "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "success": false, "error": "Forbidden", "code": "FORBIDDEN" } } }
      },
...
      "ErrorConflict": {
        "description": "Conflict",
        "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "success": false, "error": "Conflict", "code": "CONFLICT" } } }
      },
      "ErrorInternal": {
        "description": "Internal Server Error",
        "headers": { "X-API-Version": { "$ref": "#/components/headers/XApiVersion" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "success": false, "error": "Internal server error", "code": "INTERNAL_ERROR" } } }
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "message": { "type": "string" },
          "data": {}
        },
        "additionalProperties": true
      },
      "PaginatedSuccessResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "data": {},
          "pagination": {
            "type": "object",
            "properties": {
              "page": { "type": "integer" },
              "pageSize": { "type": "integer" },
              "total": { "type": "integer" },
              "totalPages": { "type": "integer" }
            },
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error", "code"],
        "properties": {
          "success": { "type": "boolean", "const": false },
          "error": { "type": "string" },
          "code": { "type": "string" }
        },
        "additionalProperties": true
      },
...
    }
  }
}